+++ NEWS.rst
@@ -56,9 +56,9 @@ Changes in 5.14 (released ??/??/2015)
 	included in the log file name now by default.
 
 *	Two new helper functions were added to :mod:`ll.misc`:
-	:func:`format_exception_type` formats the class of an exception (e.g.
-	``ValueError`` or ``http.client.HTTPException``) and :func:`format_exception`
-	formats the exception itself::
+	:func:`format_class` formats the name of a class object (e.g. ``ValueError``
+	or ``http.client.HTTPException``) and :func:`format_exception`
+	formats an exception::
 
 		>>> misc.format_exception(ValueError("bad value"))
 		'ValueError: bad value'